[Fixes #5901] Move backup and restore management commands to a custom module
