+++ geonode/base/translation.py
@@ -20,7 +20,7 @@
 
 from modeltranslation.translator import translator, TranslationOptions
 from geonode.base.models import (TopicCategory, SpatialRepresentationType, Region,
-                                 RestrictionCodeType, Backup, License, ResourceBase)
+                                 RestrictionCodeType, License, ResourceBase)
 
 
 class TopicCategoryTranslationOptions(TranslationOptions):
@@ -39,10 +39,6 @@ class RestrictionCodeTypeTranslationOptions(TranslationOptions):
     fields = ('description', 'gn_description',)
 
 
-class BackupTranslationOptions(TranslationOptions):
-    fields = ('name', 'description',)
-
-
 class LicenseTranslationOptions(TranslationOptions):
     fields = ('name', 'description', 'license_text',)
 
@@ -51,6 +47,5 @@ translator.register(TopicCategory, TopicCategoryTranslationOptions)
 translator.register(SpatialRepresentationType, SpatialRepresentationTypeTranslationOptions)
 translator.register(Region, RegionTranslationOptions)
 translator.register(RestrictionCodeType, RestrictionCodeTypeTranslationOptions)
-translator.register(Backup, BackupTranslationOptions)
 translator.register(License, LicenseTranslationOptions)
 translator.register(ResourceBase)
